Refactor environments into bits
